Implement functions to cache inverse matrix
Provide implementations for the two functions makeCacheMatrix
and cacheSolve.

makeCacheMatrix:  Creates a new object which holds a matrix
and optionally a cached inverse.  It will also automatically
remove the inverse when the matrix is changed.

cacheSolve:  Calculates the inverse of a matrix which was
created by the makeCacheMatrix function and takes care of
caching and using the inverse to improve performance.